feat(schema-registry): Mini refactoring for the ProducerRecord part
